+++ Changelog.md
@@ -5,3 +5,7 @@ Version 1.0.0
 
 Version 1.0.1
 - Fixed /js dofile throwing a error when you don't specify a filename
+
+Version 1.1.0
+- Added JSDoc tag, it will in the near future be readable in game but for now it will be used to document each function
+- Added JSDoc tags for ScriptingGlobals